+++ .gitignore
@@ -23,6 +23,3 @@ bin/
 # fabric
 
 run/
-
-# spinnery
-src/main/java/spinnery/debug
