+++ docs/constant-values.html
@@ -135,7 +135,7 @@
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colFirst"><a name="spinnery.common.container.BaseContainer.PLAYER_INVENTORY">
+<td class="colFirst"><a name="spinnery.common.handler.BaseScreenHandler.PLAYER_INVENTORY">
 <!--   -->
 </a><code>public&nbsp;static&nbsp;final&nbsp;int</code></td>
 <td><code><a href="spinnery/common/BaseContainer.html#PLAYER_INVENTORY">PLAYER_INVENTORY</a></code></td>
