+++ docs/deprecated-list.html
@@ -88,34 +88,34 @@
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colOne"><a href="spinnery/common/BaseContainer.html#addSlot-net.minecraft.container.Slot-">spinnery.common.container.BaseContainer.addSlot(Slot)</a></td>
+<td class="colOne"><a href="spinnery/common/BaseContainer.html#addSlot-net.minecraft.container.Slot-">spinnery.common.handler.BaseScreenHandler.addSlot(Slot)</a></td>
 </tr>
 <tr class="rowColor">
-<td class="colOne"><a href="spinnery/common/BaseContainer.html#canUse-net.minecraft.entity.player.PlayerEntity-">spinnery.common.container.BaseContainer.canUse(PlayerEntity)</a></td>
+<td class="colOne"><a href="spinnery/common/BaseContainer.html#canUse-net.minecraft.entity.player.PlayerEntity-">spinnery.common.handler.BaseScreenHandler.canUse(PlayerEntity)</a></td>
 </tr>
 <tr class="altColor">
-<td class="colOne"><a href="spinnery/common/BaseContainer.html#clearCraftingSlots--">spinnery.common.container.BaseContainer.clearCraftingSlots()</a></td>
+<td class="colOne"><a href="spinnery/common/BaseContainer.html#clearCraftingSlots--">spinnery.common.handler.BaseScreenHandler.clearCraftingSlots()</a></td>
 </tr>
 <tr class="rowColor">
-<td class="colOne"><a href="spinnery/common/BaseContainer.html#getCraftingHeight--">spinnery.common.container.BaseContainer.getCraftingHeight()</a></td>
+<td class="colOne"><a href="spinnery/common/BaseContainer.html#getCraftingHeight--">spinnery.common.handler.BaseScreenHandler.getCraftingHeight()</a></td>
 </tr>
 <tr class="altColor">
-<td class="colOne"><a href="spinnery/common/BaseContainer.html#getCraftingResultSlotIndex--">spinnery.common.container.BaseContainer.getCraftingResultSlotIndex()</a></td>
+<td class="colOne"><a href="spinnery/common/BaseContainer.html#getCraftingResultSlotIndex--">spinnery.common.handler.BaseScreenHandler.getCraftingResultSlotIndex()</a></td>
 </tr>
 <tr class="rowColor">
-<td class="colOne"><a href="spinnery/common/BaseContainer.html#getCraftingSlotCount--">spinnery.common.container.BaseContainer.getCraftingSlotCount()</a></td>
+<td class="colOne"><a href="spinnery/common/BaseContainer.html#getCraftingSlotCount--">spinnery.common.handler.BaseScreenHandler.getCraftingSlotCount()</a></td>
 </tr>
 <tr class="altColor">
-<td class="colOne"><a href="spinnery/common/BaseContainer.html#getCraftingWidth--">spinnery.common.container.BaseContainer.getCraftingWidth()</a></td>
+<td class="colOne"><a href="spinnery/common/BaseContainer.html#getCraftingWidth--">spinnery.common.handler.BaseScreenHandler.getCraftingWidth()</a></td>
 </tr>
 <tr class="rowColor">
-<td class="colOne"><a href="spinnery/common/BaseContainer.html#matches-net.minecraft.recipe.Recipe-">spinnery.common.container.BaseContainer.matches(Recipe&lt;? super Inventory&gt;)</a></td>
+<td class="colOne"><a href="spinnery/common/BaseContainer.html#matches-net.minecraft.recipe.Recipe-">spinnery.common.handler.BaseScreenHandler.matches(Recipe&lt;? super Inventory&gt;)</a></td>
 </tr>
 <tr class="altColor">
-<td class="colOne"><a href="spinnery/common/BaseContainer.html#onSlotClick-int-int-net.minecraft.container.SlotActionType-net.minecraft.entity.player.PlayerEntity-">spinnery.common.container.BaseContainer.onSlotClick(int, int, SlotActionType, PlayerEntity)</a></td>
+<td class="colOne"><a href="spinnery/common/BaseContainer.html#onSlotClick-int-int-net.minecraft.container.SlotActionType-net.minecraft.entity.player.PlayerEntity-">spinnery.common.handler.BaseScreenHandler.onSlotClick(int, int, SlotActionType, PlayerEntity)</a></td>
 </tr>
 <tr class="rowColor">
-<td class="colOne"><a href="spinnery/common/BaseContainer.html#populateRecipeFinder-net.minecraft.recipe.RecipeFinder-">spinnery.common.container.BaseContainer.populateRecipeFinder(RecipeFinder)</a></td>
+<td class="colOne"><a href="spinnery/common/BaseContainer.html#populateRecipeFinder-net.minecraft.recipe.RecipeFinder-">spinnery.common.handler.BaseScreenHandler.populateRecipeFinder(RecipeFinder)</a></td>
 </tr>
 </tbody>
 </table>
