+++ docs/index-files/index-16.html
@@ -96,7 +96,7 @@
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../spinnery/widget/WInterface.html#setClientside-java.lang.Boolean-">setClientside(Boolean)</a></span> - Method in class spinnery.widget.<a href="../spinnery/widget/WInterface.html" title="class in spinnery.widget">WInterface</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../spinnery/widget/WInterface.html#setContainer-spinnery.common.container.BaseContainer-">setContainer(BaseContainer)</a></span> - Method in class spinnery.widget.<a href="../spinnery/widget/WInterface.html" title="class in spinnery.widget">WInterface</a></dt>
+<dt><span class="memberNameLink"><a href="../spinnery/widget/WInterface.html#setContainer-spinnery.common.handler.BaseScreenHandler-">setContainer(BaseContainer)</a></span> - Method in class spinnery.widget.<a href="../spinnery/widget/WInterface.html" title="class in spinnery.widget">WInterface</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../spinnery/widget/WDynamicImage.html#setCurrentImage-int-">setCurrentImage(int)</a></span> - Method in class spinnery.widget.<a href="../spinnery/widget/WDynamicImage.html" title="class in spinnery.widget">WDynamicImage</a></dt>
 <dd>&nbsp;</dd>
