+++ docs/index-files/index-20.html
@@ -116,13 +116,13 @@
 <dd>&nbsp;</dd>
 <dt><a href="../spinnery/widget/WInterface.html" title="class in spinnery.widget"><span class="typeNameLink">WInterface</span></a> - Class in <a href="../spinnery/widget/package-summary.html">spinnery.widget</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../spinnery/widget/WInterface.html#WInterface-spinnery.common.container.BaseContainer-">WInterface(BaseContainer)</a></span> - Constructor for class spinnery.widget.<a href="../spinnery/widget/WInterface.html" title="class in spinnery.widget">WInterface</a></dt>
+<dt><span class="memberNameLink"><a href="../spinnery/widget/WInterface.html#WInterface-spinnery.common.handler.BaseScreenHandler-">WInterface(BaseContainer)</a></span> - Constructor for class spinnery.widget.<a href="../spinnery/widget/WInterface.html" title="class in spinnery.widget">WInterface</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../spinnery/widget/WInterface.html#WInterface-spinnery.widget.WPosition-">WInterface(WPosition)</a></span> - Constructor for class spinnery.widget.<a href="../spinnery/widget/WInterface.html" title="class in spinnery.widget">WInterface</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../spinnery/widget/WInterface.html#WInterface-spinnery.widget.WPosition-spinnery.widget.WSize-">WInterface(WPosition, WSize)</a></span> - Constructor for class spinnery.widget.<a href="../spinnery/widget/WInterface.html" title="class in spinnery.widget">WInterface</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../spinnery/widget/WInterface.html#WInterface-spinnery.widget.WPosition-spinnery.widget.WSize-spinnery.common.container.BaseContainer-">WInterface(WPosition, WSize, BaseContainer)</a></span> - Constructor for class spinnery.widget.<a href="../spinnery/widget/WInterface.html" title="class in spinnery.widget">WInterface</a></dt>
+<dt><span class="memberNameLink"><a href="../spinnery/widget/WInterface.html#WInterface-spinnery.widget.WPosition-spinnery.widget.WSize-spinnery.common.handler.BaseScreenHandler-">WInterface(WPosition, WSize, BaseContainer)</a></span> - Constructor for class spinnery.widget.<a href="../spinnery/widget/WInterface.html" title="class in spinnery.widget">WInterface</a></dt>
 <dd>&nbsp;</dd>
 <dt><a href="../spinnery/widget/WInterfaceHolder.html" title="class in spinnery.widget"><span class="typeNameLink">WInterfaceHolder</span></a> - Class in <a href="../spinnery/widget/package-summary.html">spinnery.widget</a></dt>
 <dd>&nbsp;</dd>
