+++ docs/spinnery/common/BaseContainer.html
@@ -106,7 +106,7 @@ var activeTableTab = "activeTableTab";
 <li>net.minecraft.container.CraftingContainer&lt;net.minecraft.inventory.Inventory&gt;</li>
 <li>
 <ul class="inheritance">
-<li>spinnery.common.container.BaseContainer</li>
+<li>spinnery.common.handler.BaseScreenHandler</li>
 </ul>
 </li>
 </ul>
@@ -377,7 +377,7 @@ implements net.minecraft.util.Tickable</pre>
 <pre>public static final&nbsp;int PLAYER_INVENTORY</pre>
 <dl>
 <dt><span class="seeLabel">See Also:</span></dt>
-<dd><a href="../../constant-values.html#spinnery.common.container.BaseContainer.PLAYER_INVENTORY">Constant Field Values</a></dd>
+<dd><a href="../../constant-values.html#spinnery.common.handler.BaseScreenHandler.PLAYER_INVENTORY">Constant Field Values</a></dd>
 </dl>
 </li>
 </ul>
