+++ docs/spinnery/common/BaseContainerScreen.html
@@ -112,7 +112,7 @@ var activeTableTab = "activeTableTab";
 <li>net.minecraft.client.gui.screen.ingame.AbstractContainerScreen&lt;T&gt;</li>
 <li>
 <ul class="inheritance">
-<li>spinnery.client.screen.BaseContainerScreen&lt;T&gt;</li>
+<li>spinnery.client.screen.BaseHandledScreen&lt;T&gt;</li>
 </ul>
 </li>
 </ul>
@@ -438,7 +438,7 @@ extends net.minecraft.client.gui.screen.ingame.AbstractContainerScreen&lt;T&gt;<
 <!--   -->
 </a>
 <h3>Constructor Detail</h3>
-<a name="BaseContainerScreen-net.minecraft.text.Text-spinnery.common.container.BaseContainer-net.minecraft.entity.player.PlayerEntity-">
+<a name="BaseContainerScreen-net.minecraft.text.Text-spinnery.common.handler.BaseScreenHandler-net.minecraft.entity.player.PlayerEntity-">
 <!--   -->
 </a><a name="BaseContainerScreen-net.minecraft.text.Text-T-net.minecraft.entity.player.PlayerEntity-">
 <!--   -->
