+++ docs/spinnery/common/class-use/BaseContainer.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (1.8.0_211) on Thu Jan 23 20:25:11 BRST 2020 -->
-<title>Uses of Class spinnery.common.container.BaseContainer</title>
+<title>Uses of Class spinnery.common.handler.BaseScreenHandler</title>
 <meta name="date" content="2020-01-23">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../script.js"></script>
@@ -12,7 +12,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="Uses of Class spinnery.common.container.BaseContainer";
+            parent.document.title="Uses of Class spinnery.common.handler.BaseScreenHandler";
         }
     }
     catch(err) {
@@ -70,7 +70,7 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h2 title="Uses of Class spinnery.common.container.BaseContainer" class="title">Uses of Class<br>spinnery.common.container.BaseContainer</h2>
+<h2 title="Uses of Class spinnery.common.handler.BaseScreenHandler" class="title">Uses of Class<br>spinnery.common.handler.BaseScreenHandler</h2>
 </div>
 <div class="classUseContainer">
 <ul class="blockList">
@@ -174,7 +174,7 @@
 <tbody>
 <tr class="altColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><span class="typeNameLabel">WInterface.</span><code><span class="memberNameLink"><a href="../../../spinnery/widget/WInterface.html#setContainer-spinnery.common.container.BaseContainer-">setContainer</a></span>(<a href="../../../spinnery/common/BaseContainer.html" title="class in spinnery.common">BaseContainer</a>&nbsp;linkedContainer)</code>&nbsp;</td>
+<td class="colLast"><span class="typeNameLabel">WInterface.</span><code><span class="memberNameLink"><a href="../../../spinnery/widget/WInterface.html#setContainer-spinnery.common.handler.BaseScreenHandler-">setContainer</a></span>(<a href="../../../spinnery/common/BaseContainer.html" title="class in spinnery.common">BaseContainer</a>&nbsp;linkedContainer)</code>&nbsp;</td>
 </tr>
 </tbody>
 </table>
@@ -185,10 +185,10 @@
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../spinnery/widget/WInterface.html#WInterface-spinnery.common.container.BaseContainer-">WInterface</a></span>(<a href="../../../spinnery/common/BaseContainer.html" title="class in spinnery.common">BaseContainer</a>&nbsp;linkedContainer)</code>&nbsp;</td>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../spinnery/widget/WInterface.html#WInterface-spinnery.common.handler.BaseScreenHandler-">WInterface</a></span>(<a href="../../../spinnery/common/BaseContainer.html" title="class in spinnery.common">BaseContainer</a>&nbsp;linkedContainer)</code>&nbsp;</td>
 </tr>
 <tr class="rowColor">
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../spinnery/widget/WInterface.html#WInterface-spinnery.widget.WPosition-spinnery.widget.WSize-spinnery.common.container.BaseContainer-">WInterface</a></span>(<a href="../../../spinnery/widget/WPosition.html" title="class in spinnery.widget">WPosition</a>&nbsp;position,
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../spinnery/widget/WInterface.html#WInterface-spinnery.widget.WPosition-spinnery.widget.WSize-spinnery.common.handler.BaseScreenHandler-">WInterface</a></span>(<a href="../../../spinnery/widget/WPosition.html" title="class in spinnery.widget">WPosition</a>&nbsp;position,
           <a href="../../../spinnery/widget/WSize.html" title="class in spinnery.widget">WSize</a>&nbsp;size,
           <a href="../../../spinnery/common/BaseContainer.html" title="class in spinnery.common">BaseContainer</a>&nbsp;linkedContainer)</code>&nbsp;</td>
 </tr>
