+++ docs/spinnery/common/class-use/BaseContainerScreen.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (1.8.0_211) on Thu Jan 23 20:25:11 BRST 2020 -->
-<title>Uses of Class spinnery.client.screen.BaseContainerScreen</title>
+<title>Uses of Class spinnery.client.screen.BaseHandledScreen</title>
 <meta name="date" content="2020-01-23">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../script.js"></script>
@@ -12,7 +12,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="Uses of Class spinnery.client.screen.BaseContainerScreen";
+            parent.document.title="Uses of Class spinnery.client.screen.BaseHandledScreen";
         }
     }
     catch(err) {
@@ -70,9 +70,9 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h2 title="Uses of Class spinnery.client.screen.BaseContainerScreen" class="title">Uses of Class<br>spinnery.client.screen.BaseContainerScreen</h2>
+<h2 title="Uses of Class spinnery.client.screen.BaseHandledScreen" class="title">Uses of Class<br>spinnery.client.screen.BaseHandledScreen</h2>
 </div>
-<div class="classUseContainer">No usage of spinnery.client.screen.BaseContainerScreen</div>
+<div class="classUseContainer">No usage of spinnery.client.screen.BaseHandledScreen</div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
 <div class="bottomNav"><a name="navbar.bottom">
 <!--   -->
