+++ docs/spinnery/debug/TestContainer.html
@@ -69,9 +69,9 @@
 <ul class="subNavList">
 <li>Summary:&nbsp;</li>
 <li>Nested&nbsp;|&nbsp;</li>
-<li><a href="#fields.inherited.from.class.spinnery.common.container.BaseContainer">Field</a>&nbsp;|&nbsp;</li>
+<li><a href="#fields.inherited.from.class.spinnery.common.handler.BaseScreenHandler">Field</a>&nbsp;|&nbsp;</li>
 <li><a href="#constructor.summary">Constr</a>&nbsp;|&nbsp;</li>
-<li><a href="#methods.inherited.from.class.spinnery.common.container.BaseContainer">Method</a></li>
+<li><a href="#methods.inherited.from.class.spinnery.common.handler.BaseScreenHandler">Method</a></li>
 </ul>
 <ul class="subNavList">
 <li>Detail:&nbsp;</li>
@@ -100,7 +100,7 @@
 <li>net.minecraft.container.CraftingContainer&lt;net.minecraft.inventory.Inventory&gt;</li>
 <li>
 <ul class="inheritance">
-<li><a href="../../spinnery/common/BaseContainer.html" title="class in spinnery.common">spinnery.common.container.BaseContainer</a></li>
+<li><a href="../../spinnery/common/BaseContainer.html" title="class in spinnery.common">spinnery.common.handler.BaseScreenHandler</a></li>
 <li>
 <ul class="inheritance">
 <li>spinnery.debug.TestContainer</li>
@@ -137,7 +137,7 @@ extends <a href="../../spinnery/common/BaseContainer.html" title="class in spinn
 </a>
 <h3>Field Summary</h3>
 <ul class="blockList">
-<li class="blockList"><a name="fields.inherited.from.class.spinnery.common.container.BaseContainer">
+<li class="blockList"><a name="fields.inherited.from.class.spinnery.common.handler.BaseScreenHandler">
 <!--   -->
 </a>
 <h3>Fields inherited from class&nbsp;spinnery.common.<a href="../../spinnery/common/BaseContainer.html" title="class in spinnery.common">BaseContainer</a></h3>
@@ -177,7 +177,7 @@ extends <a href="../../spinnery/common/BaseContainer.html" title="class in spinn
 </a>
 <h3>Method Summary</h3>
 <ul class="blockList">
-<li class="blockList"><a name="methods.inherited.from.class.spinnery.common.container.BaseContainer">
+<li class="blockList"><a name="methods.inherited.from.class.spinnery.common.handler.BaseScreenHandler">
 <!--   -->
 </a>
 <h3>Methods inherited from class&nbsp;spinnery.common.<a href="../../spinnery/common/BaseContainer.html" title="class in spinnery.common">BaseContainer</a></h3>
@@ -282,9 +282,9 @@ extends <a href="../../spinnery/common/BaseContainer.html" title="class in spinn
 <ul class="subNavList">
 <li>Summary:&nbsp;</li>
 <li>Nested&nbsp;|&nbsp;</li>
-<li><a href="#fields.inherited.from.class.spinnery.common.container.BaseContainer">Field</a>&nbsp;|&nbsp;</li>
+<li><a href="#fields.inherited.from.class.spinnery.common.handler.BaseScreenHandler">Field</a>&nbsp;|&nbsp;</li>
 <li><a href="#constructor.summary">Constr</a>&nbsp;|&nbsp;</li>
-<li><a href="#methods.inherited.from.class.spinnery.common.container.BaseContainer">Method</a></li>
+<li><a href="#methods.inherited.from.class.spinnery.common.handler.BaseScreenHandler">Method</a></li>
 </ul>
 <ul class="subNavList">
 <li>Detail:&nbsp;</li>
