+++ docs/spinnery/widget/WInterface.html
@@ -211,7 +211,7 @@ extends <a href="../../spinnery/widget/WWidget.html" title="class in spinnery.wi
 <th class="colOne" scope="col">Constructor and Description</th>
 </tr>
 <tr class="altColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../spinnery/widget/WInterface.html#WInterface-spinnery.common.container.BaseContainer-">WInterface</a></span>(<a href="../../spinnery/common/BaseContainer.html" title="class in spinnery.common">BaseContainer</a>&nbsp;linkedContainer)</code>&nbsp;</td>
+<td class="colOne"><code><span class="memberNameLink"><a href="../../spinnery/widget/WInterface.html#WInterface-spinnery.common.handler.BaseScreenHandler-">WInterface</a></span>(<a href="../../spinnery/common/BaseContainer.html" title="class in spinnery.common">BaseContainer</a>&nbsp;linkedContainer)</code>&nbsp;</td>
 </tr>
 <tr class="rowColor">
 <td class="colOne"><code><span class="memberNameLink"><a href="../../spinnery/widget/WInterface.html#WInterface-spinnery.widget.WPosition-">WInterface</a></span>(<a href="../../spinnery/widget/WPosition.html" title="class in spinnery.widget">WPosition</a>&nbsp;position)</code>&nbsp;</td>
@@ -221,7 +221,7 @@ extends <a href="../../spinnery/widget/WWidget.html" title="class in spinnery.wi
           <a href="../../spinnery/widget/WSize.html" title="class in spinnery.widget">WSize</a>&nbsp;size)</code>&nbsp;</td>
 </tr>
 <tr class="rowColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../spinnery/widget/WInterface.html#WInterface-spinnery.widget.WPosition-spinnery.widget.WSize-spinnery.common.container.BaseContainer-">WInterface</a></span>(<a href="../../spinnery/widget/WPosition.html" title="class in spinnery.widget">WPosition</a>&nbsp;position,
+<td class="colOne"><code><span class="memberNameLink"><a href="../../spinnery/widget/WInterface.html#WInterface-spinnery.widget.WPosition-spinnery.widget.WSize-spinnery.common.handler.BaseScreenHandler-">WInterface</a></span>(<a href="../../spinnery/widget/WPosition.html" title="class in spinnery.widget">WPosition</a>&nbsp;position,
           <a href="../../spinnery/widget/WSize.html" title="class in spinnery.widget">WSize</a>&nbsp;size,
           <a href="../../spinnery/common/BaseContainer.html" title="class in spinnery.common">BaseContainer</a>&nbsp;linkedContainer)</code>&nbsp;</td>
 </tr>
@@ -294,7 +294,7 @@ extends <a href="../../spinnery/widget/WWidget.html" title="class in spinnery.wi
 </tr>
 <tr id="i13" class="rowColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../spinnery/widget/WInterface.html#setContainer-spinnery.common.container.BaseContainer-">setContainer</a></span>(<a href="../../spinnery/common/BaseContainer.html" title="class in spinnery.common">BaseContainer</a>&nbsp;linkedContainer)</code>&nbsp;</td>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../spinnery/widget/WInterface.html#setContainer-spinnery.common.handler.BaseScreenHandler-">setContainer</a></span>(<a href="../../spinnery/common/BaseContainer.html" title="class in spinnery.common">BaseContainer</a>&nbsp;linkedContainer)</code>&nbsp;</td>
 </tr>
 <tr id="i14" class="altColor">
 <td class="colFirst"><code>void</code></td>
@@ -442,7 +442,7 @@ extends <a href="../../spinnery/widget/WWidget.html" title="class in spinnery.wi
 <!--   -->
 </a>
 <h3>Constructor Detail</h3>
-<a name="WInterface-spinnery.common.container.BaseContainer-">
+<a name="WInterface-spinnery.common.handler.BaseScreenHandler-">
 <!--   -->
 </a>
 <ul class="blockList">
@@ -470,7 +470,7 @@ extends <a href="../../spinnery/widget/WWidget.html" title="class in spinnery.wi
                   <a href="../../spinnery/widget/WSize.html" title="class in spinnery.widget">WSize</a>&nbsp;size)</pre>
 </li>
 </ul>
-<a name="WInterface-spinnery.widget.WPosition-spinnery.widget.WSize-spinnery.common.container.BaseContainer-">
+<a name="WInterface-spinnery.widget.WPosition-spinnery.widget.WSize-spinnery.common.handler.BaseScreenHandler-">
 <!--   -->
 </a>
 <ul class="blockListLast">
@@ -561,7 +561,7 @@ extends <a href="../../spinnery/widget/WWidget.html" title="class in spinnery.wi
 <pre>public&nbsp;<a href="../../spinnery/common/BaseContainer.html" title="class in spinnery.common">BaseContainer</a>&nbsp;getContainer()</pre>
 </li>
 </ul>
-<a name="setContainer-spinnery.common.container.BaseContainer-">
+<a name="setContainer-spinnery.common.handler.BaseScreenHandler-">
 <!--   -->
 </a>
 <ul class="blockList">
