+++ docs/spinnery/widget/class-use/WPosition.html
@@ -201,7 +201,7 @@
           <a href="../../../spinnery/widget/WSize.html" title="class in spinnery.widget">WSize</a>&nbsp;size)</code>&nbsp;</td>
 </tr>
 <tr class="rowColor">
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../spinnery/widget/WInterface.html#WInterface-spinnery.widget.WPosition-spinnery.widget.WSize-spinnery.common.container.BaseContainer-">WInterface</a></span>(<a href="../../../spinnery/widget/WPosition.html" title="class in spinnery.widget">WPosition</a>&nbsp;position,
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../spinnery/widget/WInterface.html#WInterface-spinnery.widget.WPosition-spinnery.widget.WSize-spinnery.common.handler.BaseScreenHandler-">WInterface</a></span>(<a href="../../../spinnery/widget/WPosition.html" title="class in spinnery.widget">WPosition</a>&nbsp;position,
           <a href="../../../spinnery/widget/WSize.html" title="class in spinnery.widget">WSize</a>&nbsp;size,
           <a href="../../../spinnery/common/BaseContainer.html" title="class in spinnery.common">BaseContainer</a>&nbsp;linkedContainer)</code>&nbsp;</td>
 </tr>
