+++ gradle.properties
@@ -8,7 +8,7 @@ loader_version = 0.8.8+build.202
 # Mod Properties
 mod_name = spinnery
 mod_group = com.github.vini2003
-mod_version = 3.0.
+mod_version = 3.0.55
 mod_target = fabric-1.16.x
 
 # Dependencies
