+++ src/main/java/spinnery/Spinnery.java
@@ -1,10 +1,13 @@
 package spinnery;
 
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.loader.api.FabricLoader;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import spinnery.common.configuration.registry.ConfigurationRegistry;
 import spinnery.common.registry.NetworkRegistry;
+import spinnery.debug.DebugCommands;
+import spinnery.debug.DebugScreenHandlers;
 
 public class Spinnery implements ModInitializer {
 	public static final String LOG_ID = "Spinnery";
@@ -15,5 +18,10 @@ public class Spinnery implements ModInitializer {
 	public void onInitialize() {
 		NetworkRegistry.initialize();
 		ConfigurationRegistry.initialize();
+
+		if (FabricLoader.getInstance().isDevelopmentEnvironment()) {
+			DebugScreenHandlers.initialize();
+			DebugCommands.initialize();
+		}
 	}
 }
