+++ src/main/java/spinnery/SpinneryClient.java
@@ -1,10 +1,14 @@
 package spinnery;
 
 import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.loader.api.FabricLoader;
 import spinnery.client.integration.SpinneryConfigurationScreen;
 import spinnery.common.registry.NetworkRegistry;
 import spinnery.common.registry.ThemeResourceRegistry;
 import spinnery.common.registry.WidgetRegistry;
+import spinnery.debug.DebugHandledScreen;
+import spinnery.debug.DebugHandledScreens;
+import spinnery.debug.DebugScreenHandlers;
 
 public class SpinneryClient implements ClientModInitializer {
 	@Override
@@ -14,5 +18,9 @@ public class SpinneryClient implements ClientModInitializer {
 		ThemeResourceRegistry.initialize();
 
 		SpinneryConfigurationScreen.initialize();
+
+		if (FabricLoader.getInstance().isDevelopmentEnvironment()) {
+			DebugHandledScreens.initialize();
+		}
 	}
 }
