+++ src/main/java/spinnery/client/screen/BaseScreen.java
@@ -21,6 +21,8 @@ public class BaseScreen extends Screen implements WInterfaceProvider {
 
 	@Override
 	public void render(MatrixStack matrices, int mouseX, int mouseY, float tick) {
+		this.fillGradient(matrices, 0, 0, this.width, this.height, -1072689136, -804253680);
+
 		VertexConsumerProvider.Immediate provider = MinecraftClient.getInstance().getBufferBuilders().getEffectVertexConsumers();
 
 		getInterface().draw(matrices, provider);
