+++ src/main/java/spinnery/common/registry/NetworkRegistry.java
@@ -9,7 +9,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.PacketByteBuf;
 import net.minecraft.util.Identifier;
-import spinnery.common.container.BaseContainer;
+import spinnery.common.handler.BaseScreenHandler;
 import spinnery.widget.WSlot;
 import spinnery.widget.api.Action;
 import spinnery.widget.api.WNetworked;
@@ -180,8 +180,8 @@ public class NetworkRegistry {
 			Action action = packetByteBuffer.readEnumConstant(Action.class);
 
 			packetContext.getTaskQueue().execute(() -> {
-				if (packetContext.getPlayer().currentScreenHandler instanceof BaseContainer && packetContext.getPlayer().currentScreenHandler.syncId == syncId) {
-					((BaseContainer) packetContext.getPlayer().currentScreenHandler).onSlotAction(slotNumber, inventoryNumber, button, action, packetContext.getPlayer());
+				if (packetContext.getPlayer().currentScreenHandler instanceof BaseScreenHandler && packetContext.getPlayer().currentScreenHandler.syncId == syncId) {
+					((BaseScreenHandler) packetContext.getPlayer().currentScreenHandler).onSlotAction(slotNumber, inventoryNumber, button, action, packetContext.getPlayer());
 				}
 			});
 		});
@@ -193,8 +193,8 @@ public class NetworkRegistry {
 			Action action = packetByteBuffer.readEnumConstant(Action.class);
 
 			packetContext.getTaskQueue().execute(() -> {
-				if (packetContext.getPlayer().currentScreenHandler instanceof BaseContainer && packetContext.getPlayer().currentScreenHandler.syncId == syncId) {
-					((BaseContainer) packetContext.getPlayer().currentScreenHandler).onSlotDrag(slotNumbers, inventoryNumbers, action);
+				if (packetContext.getPlayer().currentScreenHandler instanceof BaseScreenHandler && packetContext.getPlayer().currentScreenHandler.syncId == syncId) {
+					((BaseScreenHandler) packetContext.getPlayer().currentScreenHandler).onSlotDrag(slotNumbers, inventoryNumbers, action);
 				}
 			});
 		}));
@@ -204,8 +204,8 @@ public class NetworkRegistry {
 			WNetworked.Event event = packetByteBuf.readEnumConstant(WNetworked.Event.class);
 			CompoundTag payload = packetByteBuf.readCompoundTag();
 			packetContext.getTaskQueue().execute(() -> {
-				if (packetContext.getPlayer().currentScreenHandler instanceof BaseContainer) {
-					((BaseContainer) packetContext.getPlayer().currentScreenHandler).onInterfaceEvent(widgetSyncId, event, payload);
+				if (packetContext.getPlayer().currentScreenHandler instanceof BaseScreenHandler) {
+					((BaseScreenHandler) packetContext.getPlayer().currentScreenHandler).onInterfaceEvent(widgetSyncId, event, payload);
 				}
 			});
 		});
@@ -221,15 +221,15 @@ public class NetworkRegistry {
 					ItemStack stack = ItemStack.fromTag(tag);
 
 					packetContext.getTaskQueue().execute(() -> {
-						if (packetContext.getPlayer().currentScreenHandler instanceof BaseContainer && packetContext.getPlayer().currentScreenHandler.syncId == syncId) {
-							BaseContainer container = (BaseContainer) packetContext.getPlayer().currentScreenHandler;
+						if (packetContext.getPlayer().currentScreenHandler instanceof BaseScreenHandler && packetContext.getPlayer().currentScreenHandler.syncId == syncId) {
+							BaseScreenHandler handler = (BaseScreenHandler) packetContext.getPlayer().currentScreenHandler;
 
-							container.getInventory(inventoryNumber).setStack(slotNumber, stack);
+							handler.getInventory(inventoryNumber).setStack(slotNumber, stack);
 
-							WSlot slot = container.getInterface().getSlot(inventoryNumber, slotNumber);
+							WSlot slot = handler.getInterface().getSlot(inventoryNumber, slotNumber);
 
 							if (slot != null) {
-								slot.setStack(container.getInventory(inventoryNumber).getStack(slotNumber));
+								slot.setStack(handler.getInventory(inventoryNumber).getStack(slotNumber));
 							}
 						}
 					});
