+++ src/main/java/spinnery/widget/WHorizontalBar.java
@@ -28,17 +28,17 @@ public class WHorizontalBar extends WAbstractBar {
 
 		float sBGX = (int) (((sX / limit.getValue().intValue()) * progress.getValue().intValue()));
 
-		ScissorArea scissorArea = new ScissorArea((int) (x * scale), (int) (rawHeight - ((y + sY) * scale)), (int) (sX * scale), (int) (sY * scale));
+		ScissorArea area = new ScissorArea(provider, (int) (x * scale), (int) (rawHeight - ((y + sY) * scale)), (int) (sX * scale), (int) (sY * scale));
 
 		BaseRenderer.drawTexturedQuad(matrices, provider, getX(), getY(), z, getWidth(), getHeight(), getBackgroundTexture());
 
-		scissorArea.destroy();
+		area.destroy(provider);
 
-		scissorArea = new ScissorArea((int) (x * scale), (int) (rawHeight - ((y + sY) * scale)), (int) (sBGX * scale), (int) (sY * scale));
+		area = new ScissorArea(provider, (int) (x * scale), (int) (rawHeight - ((y + sY) * scale)), (int) (sBGX * scale), (int) (sY * scale));
 
 		BaseRenderer.drawTexturedQuad(matrices, provider, getX(), getY(), z, getWidth(), getHeight(), getForegroundTexture());
 
-		scissorArea.destroy();
+		area.destroy(provider);
 
 		super.draw(matrices, provider);
 	}
