+++ src/main/java/spinnery/widget/WHorizontalBoxContainer.java
@@ -165,13 +165,13 @@ public class WHorizontalBoxContainer extends WAbstractWidget implements WModifia
 			return;
 		}
 
-		ScissorArea area = new ScissorArea(this);
+		ScissorArea area = new ScissorArea(provider, this);
 
 		for (WAbstractWidget widget : widgets) {
 			widget.draw(matrices, provider);
 		}
 
-		area.destroy();
+		area.destroy(provider);
 
 		if (hasBorder()) {
 			BaseRenderer.drawQuad(matrices, provider, getX(), getY(), getZ(), getWidth(), outerBorderWidth, getStyle().asColor("border"));
